pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pi2-auth-platform"
version = "0.1.0"
description = "FastAPI migration scaffolding for the PI2 auth platform"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "asyncpg>=0.29,<1.0",
  "alembic>=1.13,<2.0",
  "pydantic-settings>=2.6,<3.0",
  "python-jose>=3.3,<4.0",
  "passlib[argon2]>=1.7,<2.0",
  "email-validator>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<10.0",
  "pytest-asyncio>=0.24,<2.0",
  "httpx>=0.27,<1.0",
  "ruff>=0.6,<1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
